Restructure Views, ViewModels & Project file
