+++ BookingManagementSystem/ViewModels/Account/LoginViewModel.cs
@@ -3,30 +3,31 @@ using CommunityToolkit.Mvvm.ComponentModel;
 using BookingManagementSystem.Core.Models;
 using BookingManagementSystem.Contracts.Services;
 using BookingManagementSystem.Core.Contracts.Services;
+using BookingManagementSystem.Core.Contracts.Repositories;
 
-namespace BookingManagementSystem.ViewModels;
+namespace BookingManagementSystem.ViewModels.Account;
 
 public partial class LoginViewModel : ObservableRecipient
 {
 
     private readonly INavigationService _navigationService;
-    private readonly IDao _dao;
+    private readonly IRepository<User> _userRepository;
 
     public IEnumerable<User> Users { get; private set; } = Enumerable.Empty<User>();
 
     public User? CurrentUser { get; private set; } // Declare as nullable
 
-    public LoginViewModel(INavigationService navigationService, IDao dao)
+    public LoginViewModel(INavigationService navigationService, IRepository<User> userRepository)
     {
         _navigationService = navigationService;
-        _dao = dao;
-        OnNavigatedTo(_dao);
+        _userRepository = userRepository;
+        OnNavigatedTo(0);
     }
 
     public async void OnNavigatedTo(object parameter)
     {
         // Load user data list
-        var users = await _dao.GetUserListDataAsync();
+        var users = await _userRepository.GetAllAsync();
         Users = users;
     }
