+++ BookingManagementSystem/ViewModels/Payment/PaymentViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 
-namespace BookingManagementSystem.ViewModels;
+namespace BookingManagementSystem.ViewModels.Payment;
 
 public partial class PaymentViewModel : ObservableRecipient
 {
