+++ BookingManagementSystem/Views/Account/LoginPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="BookingManagementSystem.Views.LoginPage"
+    x:Class="BookingManagementSystem.Views.Account.LoginPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
