+++ BookingManagementSystem/Views/Account/RecoverPasswordPage.xaml
@@ -1,18 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Page
-    x:Class="BookingManagementSystem.Views.SignupPage"
+    x:Class="BookingManagementSystem.Views.Account.RecoverPasswordPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:BookingManagementSystem.Views"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    NavigationCacheMode="Enabled"
-    CacheMode="BitmapCache">
+    mc:Ignorable="d">
+
     <Page.Background>
         <ImageBrush ImageSource="/Assets/homescreen.jpg" Stretch="Fill"/>
     </Page.Background>
-    
+
     <Grid>
         <Grid x:Name="ContentArea" Width="600" Height="450" Background="{ThemeResource AcrylicInAppFillColorBaseBrush}">
             <Grid.ColumnDefinitions>
@@ -25,18 +24,18 @@
                 <RowDefinition Height="6*"/>
                 <RowDefinition Height="*"/>
             </Grid.RowDefinitions>
+
             <Grid>
                 <Button Name="btnBackLoginPage" Click="btnBackLoginPage_Click" Background="Transparent" BorderThickness="0" Width="48" Height="48" CornerRadius="80" HorizontalAlignment="Center">
                     <SymbolIcon Symbol="Back"/>
                 </Button>
-            </Grid>
-            <StackPanel Grid.Column="1" Grid.Row="1"
-            Padding="40 0 40 30">
+            </Grid>            
+            <StackPanel Grid.Column="1" Grid.Row="1" Padding="40 0 40 30">
                 <StackPanel Orientation="Horizontal" VerticalAlignment="Center" HorizontalAlignment="Left">
                     <Image Width="30" Height="30" Source="/Assets/three-body-a.png"/>
                     <Image Source="/Assets/arrival-bio.png" Height="32" Margin="12 0 0 0" Stretch="Fill"/>
                 </StackPanel>
-                <TextBlock Padding="0 10 0 0" Text="Create account" Style="{ThemeResource TitleTextBlockStyle}"></TextBlock>
+                <TextBlock Padding="0 10 0 0" Text="Recover your account" Style="{ThemeResource TitleTextBlockStyle}"></TextBlock>
                 <TextBox Name="txtUsername" 
                  Margin="0 20 0 0" 
                  Padding="10"
@@ -49,7 +48,7 @@
                  Padding="10"
                  Height="40"
                  PasswordRevealMode="Hidden" 
-                 PlaceholderText="Password"              
+                 PlaceholderText="New Password"              
                  Background="{ThemeResource AcrylicInAppFillColorBaseBrush}"
                  AutomationProperties.Name="Password box"
                  BorderThickness="0"/>
@@ -63,20 +62,22 @@
                  AutomationProperties.Name="Password box"
                  BorderThickness="0"/>
                 <Grid Margin="0 10 0 0">
-                    <TextBlock Margin="0 8 0 0" TextWrapping="WrapWholeWords">
-                        <Hyperlink Click="Hyperlink_Login_Click">I have an account!</Hyperlink>
-                    </TextBlock>
                     <CheckBox Name="revealModeCheckBox" 
                       Content="Show password" 
                       IsChecked="False"
-                      HorizontalAlignment="Right"
+                      HorizontalAlignment="Left"
                       Checked="RevealModeCheckbox_Changed" 
                       Unchecked="RevealModeCheckbox_Changed"/>
                 </Grid>
-                <StackPanel Orientation="Vertical" Margin="0 12 0 0">
-                    <Button Name="btnSignUp" Click="btnSignUp_Click" Width="100" Height="40" HorizontalAlignment="Right" Content="Sign up" Background="#121212" Foreground="White" FontWeight="SemiBold"/>
-                </StackPanel>
+                <Grid Margin="0 0 0 0">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="*" />
+                        <ColumnDefinition Width="*" />
+                    </Grid.ColumnDefinitions>
+                    <Button Name="btnRecover" Click="btnRecover_Click" Width="100" Height="40" HorizontalAlignment="Right" Content="Recover" Background="#121212" Foreground="White" FontWeight="SemiBold" Grid.Column="1"/>
+                </Grid>
             </StackPanel>
         </Grid>
     </Grid>
+
 </Page>
