+++ BookingManagementSystem/Views/Administrator/AdministratorPage.xaml.cs
@@ -1,5 +1,5 @@
-﻿using BookingManagementSystem.ViewModels;
-
+﻿using BookingManagementSystem.ViewModels.Administrator;
+using BookingManagementSystem.Views.Client;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Media.Animation;
