+++ BookingManagementSystem/Views/Client/FAQPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="BookingManagementSystem.Views.FAQPage"
+    x:Class="BookingManagementSystem.Views.Client.FAQPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
