+++ BookingManagementSystem/Views/Client/HomePage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="BookingManagementSystem.Views.HomePage"
+    x:Class="BookingManagementSystem.Views.Client.HomePage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:BookingManagementSystem.Views"
