+++ BookingManagementSystem/Views/Client/HomePage.xaml.cs
@@ -5,7 +5,7 @@ using Microsoft.UI.Xaml.Controls;
 using Windows.Security.Cryptography;
 using Windows.Security.Cryptography.DataProtection;
 using Windows.Storage;
-using BookingManagementSystem.ViewModels;
+using BookingManagementSystem.ViewModels.Client;
 using BookingManagementSystem.Core.Models;
 using Microsoft.UI.Xaml.Media.Animation;
 using Microsoft.UI.Xaml.Media.Imaging;
@@ -14,7 +14,7 @@ using Windows.Globalization.DateTimeFormatting;
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
 
-namespace BookingManagementSystem.Views;
+namespace BookingManagementSystem.Views.Client;
 
 /// <summary>
 /// An empty page that can be used on its own or navigated to within a Frame.
