+++ BookingManagementSystem/Views/Client/RentalDetailPage.xaml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Page
-    x:Class="BookingManagementSystem.Views.RentalDetailPage"
+    x:Class="BookingManagementSystem.Views.Client.RentalDetailPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:BookingManagementSystem.Views"
